feat: jobs file cleanup option in queue manager
