+++ sushi_batch/utils/file_utils.py
@@ -1,6 +1,7 @@
 from os import path, walk
+from pathlib import Path
 
-from ..models.enums import FileTypes, Formats, Task
+from ..models.enums import FileTypes, Formats, Status, Task
 from ..models.job import Job
 
 from ..external.ffmpeg import FFmpeg
@@ -124,3 +125,30 @@ def create_jobs(src_files, dst_files, sub_files, task):
             )
         )
     return jobs
+
+
+def clean_generated_files(job_list):
+    """Delete subtitle files generated by sushi and aegisub for the specified jobs"""
+    try:
+        for job in job_list:
+            if job.status is not Status.COMPLETED: 
+                continue
+            
+            if job.task in (Task.AUDIO_SYNC_DIR, Task.AUDIO_SYNC_FIL):
+                sub_ext = Path(job.sub_file).suffix if job.sub_file else ""
+                if not sub_ext:
+                    continue
+                suffixes = [f".sushi{sub_ext}"]
+            else:
+                sub_ext = job.src_sub_ext
+                if not sub_ext:
+                    continue
+                suffixes = [f".sushi{sub_ext}", f".sushi_resampled{sub_ext}"]
+
+            for suffix in suffixes:
+                generated_file = Path(f"{job.dst_file}{suffix}")
+                generated_file.unlink(missing_ok=True)
+        cu.print_success("Generated files deleted successfully.")
+    except OSError as e:
+        cu.print_error(f"Error deleting generated files: {e}")
+
