+++ sushi_batch/utils/utils.py
@@ -3,9 +3,6 @@ import shutil
 import subprocess
 import sys
 from importlib.util import find_spec
-from pathlib import Path
-from . import console_utils as cu
-
 
 
 def is_app_installed(app_name):
@@ -36,24 +33,6 @@ def clear_logs(dirpath):
                 shutil.rmtree(entry_path)  # Recursively delete the directory
     except OSError as e:
         print(e)
-
-def clean_generated_files(job_list):
-    """Delete files generated by sushi and aegisub for the specified jobs"""
-    try:
-        for job in job_list:
-            if job.src_sub_ext is None:
-                continue
-
-            for suffix in (
-                f".sushi{job.src_sub_ext}",
-                f".sushi_resampled{job.src_sub_ext}",
-            ):
-                generated_file = Path(f"{job.dst_file}{suffix}")
-                generated_file.unlink(missing_ok=True)
-        cu.print_success("Generated files deleted successfully.")
-    except OSError as e:
-        cu.print_error(f"Error deleting generated files: {e}")
-
 
 def check_required_packages():
     _PACKAGES = ["art", "colorama", "sushi", "prettytable", "yaspin"]
